+++ R/module_session_info.R
@@ -1,6 +1,6 @@
 #' `teal` user session info module
 #'
-#' Module to display the user session info popup and teal lockfile download button
+#' Module to display the user session info popup and to download a lockfile.
 #'
 #' @rdname module_session_info
 #' @name module_session_info